requirements.txt
celery
fastapi
httpx
pydantic
python-dotenv
redis
Requests
uvicorn[standard]
